Fix #55: Merge tool providers to expose all MCP tools
Previously, the MCP server only exposed sample code search tools because
it was registering two separate tool providers (CupertinoSearchToolProvider
and SampleToolProvider), and the second registration overwrote the first.

This commit fixes the issue by:
- Creating a unified CupertinoUnifiedToolProvider that combines both
  documentation and sample code search capabilities
- Removing the separate provider implementations
- Updating ServeCommand to register only the unified provider

Now all 7 MCP tools are available:
- search_docs, list_frameworks, read_document (documentation)
- search_samples, list_samples, read_sample, read_sample_file (samples)
